% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyutils.R
\name{disconnectAutoReload}
\alias{disconnectAutoReload}
\title{Automatically reconnect shiny application when network available again}
\usage{
disconnectAutoReload()
}
\description{
Shiny application will show "Disconnected from server" pop up dialog box with a "Reload" button and grey out all the shiny application when network not available. 
This is annoying and user has to click the "Reload" button to reconnect to server.
This function helps automatically reconnect and reload application whenever network is available again. Check in 1 second period.
}
\details{
Add disconnectAutoReload() to shiny server(session,input,output)

\preformatted{

library(shiny)
library(shinyjs)
library(databunny)

ui <- fluidPage(
  titlePanel("Shiny Auto Reconnect"),
  "This shiny app will auto reconnect when ever network available again."
)

server <- function(input, output) {
disconnectAutoReload()
}

shinyApp(ui = ui, server = server)
}
}
