% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerstats.R
\name{runCPUMemoryApp}
\alias{runCPUMemoryApp}
\title{Run shiny app that display CPU and Memory usage and limit by docker container as rstudio job and display it on view pane}
\usage{
runCPUMemoryApp()
}
\description{
This is rstudio addin function, the addin will be automatically added to rstudio once packege was installed
Click Addins -> Databunny -> "Show CPU Memory Usage Limit" to activate addin
}
\details{
The backend shiny app runs on default port 56000 as a rstudio job
}
