% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerstats.R
\name{usageaddins}
\alias{usageaddins}
\title{Rstudio addin function query CPU and Memory usage and limit by docker container and show it on a viewer pane}
\usage{
usageaddins()
}
\description{
This is rstudio addin function, the addin will be automatically added to rstudio once packege was installed
Click Addins -> Databunny -> "Show CPU Memory Usage Limit" to activate addin
}
